tests/requirements.txt
psycopg2-binary==2.9.9
RESTinstance==1.3.0
requests==2.31.0
robotframework==7.0
robotframework-requests==0.9.7
robotframework-jsonlibrary==0.5
robotframework-databaselibrary==1.4.4
setuptools==69.5.1